app/requirements.txt
flask
markdown
pygments
python-slugify
GitPython
werkzeug
gunicorn
python-frontmatter
mysql-connector-python
beautifulsoup4
markupsafe
python-dotenv
